--- .env.sample
@@ -1,7 +1,0 @@
-DB_NAME=freenoms-domains
-CHROME_PATH=/usr/bin/chromium-browser
-DISCORD_WEBHOOK_ID=
-DISCORD_WEBHOOK_TOKEN=
-FREENOM_LOGIN=email
-FREENOM_PASS=password
-FREENOM_CRONJOB="0 9 * * 1"
